refactor: dedicated file for load_stac #460
This does not get rid of the problem where GeoPySparkLayerCatalog now has a dependency on GpsBatchJobs (because load_stac needs it).
